requirements.txt
django
requests
djangorestframework
django-cors-headers